+++ src/main/java/org/iesvdm/ventas/service/PedidoService.java
@@ -60,7 +60,7 @@ public class PedidoService {
 
     }
 
-    public int calcularTotal(int id){
+    public double calcularTotal(int id){
         return this.pedidoDAO.total(id);
     }
